Rewrite to make generator be a sub-package node js project instead of bun. Improved typing, etc.
This is in preparation to making the generator run as a github action, and reduce dependencies/setup requirements.

This also introduces two separate Diagnosekode subtypes for ICD-10 and ICPC-2, so that we can have code being more typesafe, avoiding accidental mixup of the two kinds of Diagnosekode. Will also probably make it easier to safely introduce ICD-11 when we want to include that.