Added search functionality to @navikt/diagnosekoder.
This adds a simple DiagnosekodeSearcher helper class to help with searching for diagnose codes. So not all consumers has to implement their own code for this.

The current implementation offers a case insensitive paged search mechanism that searches both the code and text props for given search input.

This can then be expanded on later if one wants to have more options for only searching for a code og text, etc.